+++ core/src/test/java/hudson/ClassicPluginStrategyTest.java
@@ -24,12 +24,13 @@
 package hudson;
 
 import hudson.util.VersionNumber;
+import jenkins.plugins.DetachedPluginsManager;
 import org.junit.Assert;
 import org.junit.Test;
 
 import java.util.List;
 
-import static hudson.ClassicPluginStrategy.DetachedPlugin;
+import static jenkins.plugins.DetachedPluginsManager.DetachedPlugin;
 
 /**
  * @author <a href="mailto:tom.fennelly@gmail.com">tom.fennelly@gmail.com</a>
@@ -38,20 +39,20 @@ public class ClassicPluginStrategyTest {
 
     @Test
     public void test_getDetachedPlugins() {
-        List<DetachedPlugin> list = ClassicPluginStrategy.getDetachedPlugins(new VersionNumber("1.296"));
+        List<DetachedPlugin> list = DetachedPluginsManager.getDetachedPlugins(new VersionNumber("1.296"));
 
         Assert.assertTrue(list.size() >= 14); // There were 14 at the time of writing this test
         Assert.assertNotNull(findPlugin("maven-plugin", list));
         Assert.assertNotNull(findPlugin("subversion", list));
 
         // Narrow the list to since "1.310" (the subversion detach version).
-        list = ClassicPluginStrategy.getDetachedPlugins(new VersionNumber("1.310"));
+        list = DetachedPluginsManager.getDetachedPlugins(new VersionNumber("1.310"));
         // Maven should no longer be in the list, but subversion should.
         Assert.assertNull(findPlugin("maven-plugin", list));
         Assert.assertNotNull(findPlugin("subversion", list));
 
         // Narrow the list to since "1.311" (after the subversion detach version).
-        list = ClassicPluginStrategy.getDetachedPlugins(new VersionNumber("1.311"));
+        list = DetachedPluginsManager.getDetachedPlugins(new VersionNumber("1.311"));
         // Neither Maven or subversion should be in the list.
         Assert.assertNull(findPlugin("maven-plugin", list));
         Assert.assertNull(findPlugin("subversion", list));
